+++ CHANGELOG.md
@@ -54,3 +54,5 @@ your changes, such as:
 - [public] [both] [fixed] fix config server panic caused by concurrent read and write shared object
 - [public] [both] [fixed] timezone adjust not working with apsara\_log
 - [public] [both] [added] support plugin ProcessorParseTimestampNative
+- [public] [both] [added] support plugin ProcessorParseApsaraNative
+- [public] [both] [added] support plugin ProcessorParseDelimiterNative